Homework 5 - Hypbrid List Structures
A hybrid approach that attempts to combine the advantages of both arrays (fast
indexing) and linked lists (fast inserts and removals near the start of the list) is to use a linked list
where each node stores multiple elements in a bounded array. Each node can then be considered
as a block.

Objective: 
Implement this hybrid list structure using a singly-linked list of bounded arrays

Description Given:
To traverse this list, begin at the head node and iterate through all elements of its array.
Then move to the next node and iterate through all elements of its array, and so on, until
you reach the last node, for which the next node pointer is null. See the provided at()
function for an example of a list traversal.
To insert a new element at index 𝑖, traverse the list to reach the element at index 𝑖 − 1 and
then insert the element into the same array as the element at index 𝑖 − 1. If that array is
already full, then insert a new node following the current node and copy half of the
elements from the current node into the new node, and then perform the insertion. See the
provided push_back() function for an example of an insertion at the end of the list.
To remove the element at index 𝑖, traverse the list to reach that element and then remove
it from the array. If this causes the array to now be empty, then remove the node from the
linked list.